commit grande., añadido cpu, muchas cosas, fan write etc

+++ UserInterfaceClases/HashTableWidget.py
@@ -2,6 +2,9 @@ from PyQt5.QtWidgets import QApplication, QTableWidget, QTableWidgetItem
 from PyQt5.QtCore import Qt
 import sqlite3
 from datetime import datetime
+import hash_cache
+import os
+
 class HashTableWidget(QTableWidget):
     def __init__(self):
         super().__init__()
@@ -11,20 +14,24 @@ class HashTableWidget(QTableWidget):
         self.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)    # scroll vertical solo si hace falta
         self.databaseName = "hashes.db"
 
-        data = self.getData()
+        if not os.path.exists(self.databaseName):
+            hash_cache.init_hashes_db()
+        else:
 
-        self.setColumnCount(5)
-        self.setRowCount(len(data))
+            data = self.getData()
 
-        self.setHorizontalHeaderLabels(["Hash", "State","Score", "First seen", "Last seen"])
-        self.verticalHeader().setVisible(False)
+            self.setColumnCount(5)
+            self.setRowCount(len(data))
 
-        for fila, (_, hash, state, score, fs, ls) in enumerate(data):
-            self.setItem(fila, 0, QTableWidgetItem(str(hash)))
-            self.setItem(fila, 1, QTableWidgetItem(str(state)))
-            self.setItem(fila, 2, QTableWidgetItem(str(score)))
-            self.setItem(fila, 3, QTableWidgetItem(str(datetime.fromtimestamp(fs))))
-            self.setItem(fila, 4, QTableWidgetItem(str(datetime.fromtimestamp(ls))))
+            self.setHorizontalHeaderLabels(["Hash", "State","Score", "First seen", "Last seen"])
+            self.verticalHeader().setVisible(False)
+
+            for fila, (_, hash, state, score, fs, ls) in enumerate(data):
+                self.setItem(fila, 0, QTableWidgetItem(str(hash)))
+                self.setItem(fila, 1, QTableWidgetItem(str(state)))
+                self.setItem(fila, 2, QTableWidgetItem(str(score)))
+                self.setItem(fila, 3, QTableWidgetItem(str(datetime.fromtimestamp(fs))))
+                self.setItem(fila, 4, QTableWidgetItem(str(datetime.fromtimestamp(ls))))
 
     def getData(self):
         conn = sqlite3.connect(self.databaseName)
