+++ UserInterfaceClases/InstancesTableWidget.py
@@ -1,7 +1,11 @@
 from PyQt5.QtWidgets import QApplication, QTableWidget, QTableWidgetItem
 from PyQt5.QtCore import Qt
 from datetime import datetime
+from hash_cache import *
 import sqlite3
+import os
+import hash_cache
+
 class InstancesTableWidget(QTableWidget):
     def __init__(self):
         super().__init__()
@@ -11,21 +15,24 @@ class InstancesTableWidget(QTableWidget):
         self.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)    # scroll vertical solo si hace falta
         self.databaseName = "instances.db"
 
-        data = self.getData()
+        if not os.path.exists(self.databaseName):
+            hash_cache.init_instances_db()
+        else:
+            data = self.getData()
 
-        self.setColumnCount(6)
-        self.setRowCount(len(data))
+            self.setColumnCount(6)
+            self.setRowCount(len(data))
 
-        self.setHorizontalHeaderLabels(["Hash", "Date","Pid", "Ppid", "Path", "Event"])
-        self.verticalHeader().setVisible(False)
+            self.setHorizontalHeaderLabels(["Hash", "Date","Pid", "Ppid", "Path", "Event"])
+            self.verticalHeader().setVisible(False)
 
-        for fila, (_, hash, fecha, pid, ppid, path, event) in enumerate(data):
-            self.setItem(fila, 0, QTableWidgetItem(str(hash)))
-            self.setItem(fila, 1, QTableWidgetItem(str(datetime.fromtimestamp(fecha))))
-            self.setItem(fila, 2, QTableWidgetItem(str(pid)))
-            self.setItem(fila, 3, QTableWidgetItem(str(ppid)))
-            self.setItem(fila, 4, QTableWidgetItem(str(path)))
-            self.setItem(fila, 5, QTableWidgetItem(str(event)))
+            for fila, (_, hash, fecha, pid, ppid, path, event) in enumerate(data):
+                self.setItem(fila, 0, QTableWidgetItem(str(hash)))
+                self.setItem(fila, 1, QTableWidgetItem(str(datetime.fromtimestamp(fecha))))
+                self.setItem(fila, 2, QTableWidgetItem(str(pid)))
+                self.setItem(fila, 3, QTableWidgetItem(str(ppid)))
+                self.setItem(fila, 4, QTableWidgetItem(str(path)))
+                self.setItem(fila, 5, QTableWidgetItem(str(event)))
